login and auth working, next step set cookie, models re-created, login/signup form erorrs added, ajax functions rewritten, backend new routes added

+++ react_node_express_app/src/components/carProfileComp.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import './form-error.css';
 import axios from 'axios';
+import { NavLink } from 'react-router-dom';
 
 class CarProfileComp extends React.Component{
   constructor(props){
@@ -15,7 +16,9 @@ class CarProfileComp extends React.Component{
 			colorInValid: false,
 			priceInValid: false,
 			carUpdated: false,
-			carDeleted: false
+			carDeleted: false,
+			isLoggedIn: this.props.isLoggedIn,
+			loginMsg: ""
 		}
 		this.editModeOn = this.editModeOn.bind(this)
 		this.editModeOff = this.editModeOff.bind(this)
@@ -28,23 +31,12 @@ class CarProfileComp extends React.Component{
 	}
    
 	componentDidMount(){
-		this.callApiGetSingleCar().then( res => this.setState({
-			theSingleCar: res.singleCar, 
-		})).catch( err => console.log(err));
+		this.callApiGetSingleCar()
 	}
 
-  componentDidUpdate(){
-    this.callApiGetSingleCar().then( res => this.setState({
-      theSingleCar: res.singleCar,
-    })).catch( err => console.log(err));
-  }
-
-  callApiGetSingleCar = async() => {
+  callApiGetSingleCar = () => {
     const fetchURL = '/api/cars/' + this.props.match.params.id
-    const response = await fetch(fetchURL);
-    const resBody = await response.json();
-    if (response.status !== 200) throw Error(resBody.message);
-    return resBody;
+		axios.get(fetchURL).then(res => this.setState({	theSingleCar: res.data.singleCar })).catch(err => console.log(err));
   }
 
   editModeOn = () => {
@@ -103,17 +95,19 @@ class CarProfileComp extends React.Component{
 	}
 
 	updateCarApiCall = () => {
-		const patchURL = '/api/cars/' + this.props.match.params.id
-		axios.patch(patchURL, {
-			name: this.state.editedCarName,
-			color: this.state.editedCarColor,
-			price: parseInt(this.state.editedCarPrice, 10)
-		}).then((res) => {
-			console.log(res);
-			this.callApiGetSingleCar().then( res => this.setState({
-				theSingleCar: res.singleCar,
-			})).catch( err => console.log(err))
-		}).catch(err => console.log(err))
+		if (this.state.isLoggedIn){
+			const patchURL = '/api/cars/' + this.props.match.params.id
+			axios.patch(patchURL, {
+				name: this.state.editedCarName,
+				color: this.state.editedCarColor,
+				price: parseInt(this.state.editedCarPrice, 10)
+			}).then((res) => {
+				console.log(res);
+				this.callApiGetSingleCar().then(res => this.setState({
+					theSingleCar: res.singleCar,
+				})).catch(err => console.log(err))
+			}).catch(err => console.log(err))
+		}
 	}
 
   updateCar = () => {
@@ -123,10 +117,12 @@ class CarProfileComp extends React.Component{
 	}
 
 	deleteCarApiCall = () => {
-		const deleteURL = '/api/cars/' + this.props.match.params.id
-		axios.delete(deleteURL).then( (res) => {
-			console.log(res);
-		}).catch(err => console.log(err))
+		if (this.state.isLoggedIn) {
+			const deleteURL = '/api/cars/' + this.props.match.params.id
+			axios.delete(deleteURL).then( (res) => {
+				console.log(res);
+			}).catch(err => console.log(err))
+		}
 	}
 
 	redirectToCars = () => {
@@ -141,55 +137,74 @@ class CarProfileComp extends React.Component{
   render(){
     return(
       <div className='col-md-10 col-md-offset-1'> 
-        <div className='col-sm-4'>
-					{/* {this.state.carUpdated &&
-						<div className='alert alert-success alert-dismissible' role="alert">
-							<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-							<p> Car Updated</p>
-						</div>
-					} */}
-          <h2> {this.state.theSingleCar.name} </h2>
-					<p> Color: <i className='fa fa-paint-brush'> </i> {this.state.theSingleCar.color} </p>
-					<p> Price: $ {this.state.theSingleCar.price }</p>
+				{!this.state.isLoggedIn && 
+					<div className='col-sm-6 col-sm-offset-3 mt32'> 
+						{this.state.loginMsg} <h3> <NavLink to={'/login/'}> Login </NavLink> </h3> 
+					</div>
+				}
 
-					{! this.state.isEditing && 
-						<div>
-								<button className='btn btn-success' onClick={this.editModeOn}> Edit</button>
-								<button className='btn btn-danger' onClick={this.deleteCar} >Delete</button>
-						</div>
-					}
-        </div>
-				<div className='col-sm-8'>
-					{ this.state.isEditing && 
-						<div>
-							<div className='col-sm-6'>
-								<h3> {this.state.editedCarName}</h3>
-								<p> {this.state.editedCarColor} {this.state.editedCarPrice}</p>
+				{this.state.isLoggedIn && 
+					<div className='col-sm-4'>
+						{/* {this.state.carUpdated &&
+							<div className='alert alert-success alert-dismissible' role="alert">
+								<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+								<p> Car Updated</p>
 							</div>
-							<div className='col-sm-6'>
-								<h3> Edit Car</h3>
-								<form className='form' >
-									<label htmlFor='carName'> Name</label>
-									<input type='text' className='form-control' name='carName' defaultValue={this.state.editedCarName} onChange={this.handleCarNameChange} id='carNameInput' /> 
-									{this.state.nameInValid && <span className='input-err'> ** Name Invalid</span>} <br/>
+						} */}
+						<h2> {this.state.theSingleCar.name} </h2>
+						<p> Color: <i className='fa fa-paint-brush'> </i> {this.state.theSingleCar.color} </p>
+						<p> Price: $ {this.state.theSingleCar.price }</p>
 
-									<label htmlFor='carColor'> Color</label>
-									<input type='text' className='form-control' name='carColor' defaultValue={this.state.editedCarColor} onChange={this.handleCarColorChange} id='carColorInput' /> 
-									{this.state.colorInValid && <span className='input-err'> ** Color Invalid</span>} <br/>
-
-									<label htmlFor='carPrice'> Price</label>
-									<input type='text' className='form-control' name='carPrice' defaultValue={this.state.editedCarPrice} onChange={this.handleCarPriceChange} id='carPriceInput' /> 
-									{this.state.priceInValid && <span className='input-err'> ** Price Invalid</span>}
-
-								</form>
-								<button className='btn btn-primary' type='submit' onClick={this.updateCar} id='editCarBtn'> Save</button>
-								<button className='btn btn-default' onClick ={this.editModeOff}> Close</button>
+						{! this.state.isEditing && 
+							<div>
+									<button className='btn btn-success' onClick={this.editModeOn}> Edit</button>
+									<button className='btn btn-danger' onClick={this.deleteCar} >Delete</button>
 							</div>
-						</div>
-					}
-				</div>
+						}
+					</div>
+				}
+
+				{this.state.isLoggedIn && 
+					<div className='col-sm-8'>
+						{this.state.isEditing && 
+							<div>
+								<div className='col-sm-6'>
+									<h3> {this.state.editedCarName}</h3>
+									<p> {this.state.editedCarColor} {this.state.editedCarPrice}</p>
+								</div>
+
+								<div className='col-sm-6'>
+									<h3> Edit Car</h3>
+									<form className='form'>
+										<div className='form-group'>
+											<label htmlFor='carName'> Name</label>
+											<input type='text' className='form-control' name='carName' defaultValue={this.state.editedCarName} onChange={this.handleCarNameChange} id='carNameInput' /> 
+											{this.state.nameInValid && <span className='input-err'> ** Name Invalid</span>} <br/>
+										</div>
+
+										<div className='form-group'>
+											<label htmlFor='carColor'> Color</label>
+											<input type='text' className='form-control' name='carColor' defaultValue={this.state.editedCarColor} onChange={this.handleCarColorChange} id='carColorInput' /> 
+											{this.state.colorInValid && <span className='input-err'> ** Color Invalid</span>} <br/>
+										</div>
+
+										<div className='form-group'>
+											<label htmlFor='carPrice'> Price</label>
+											<input type='number' className='form-control' name='carPrice' defaultValue={this.state.editedCarPrice} onChange={this.handleCarPriceChange} id='carPriceInput' /> 
+											{this.state.priceInValid && <span className='input-err'> ** Price Invalid</span>}
+										</div>
+									</form>
+
+									<button className='btn btn-primary' type='submit' onClick={this.updateCar} id='editCarBtn'> Save</button>
+									<button className='btn btn-default' onClick ={this.editModeOff}> Close</button>
+
+								</div>
+							</div>
+						}
+					</div>
+				}
       </div>
-    );
+    )
 	} 
 }
